format-utf-8.htm

<!DOCTYPE html>
<html>
  <head>
    <meta charset=utf-8>
    <title>EventSource: always UTF-8</title>
    <script src="/resources/testharness.js"></script>
    <script src="/resources/testharnessreport.js"></script>
  </head>
  <body>
    <div id="log"></div>
    <script>
      var test = async_test()
      test.step(function() {
        var source = new EventSource("resources/message.php?mime=text/event-stream;charset=windows-1252&message=data%3A%E2%80%A6")
        source.onmessage = function(e) {
          test.step(function() {
            assert_unreached()
            this.close()
          }, this)
          test.done()
        }
        source.onerror = function() {
          test.step(function() { assert_equals(this.readyState, this.CLOSED) }, this)
          test.done()
        }
      })
    </script>
  </body>
</html>
